move cmd_init.
